feat: add email functionality

+++ server/src/controllers/product.controller.js
@@ -6,8 +6,6 @@ const createProduct = async (req, res) => {
     const { name, description, price, category, subCategory, sizes, bestSeller } =
       req.body;
 
-    console.log(req.files);
-
     const image1 = req.files.image1 && req.files.image1[0];
     const image2 = req.files.image2 && req.files.image2[0];
     const image3 = req.files.image3 && req.files.image3[0];
@@ -22,9 +20,16 @@ const createProduct = async (req, res) => {
         });
         return result.secure_url;
       })
-    );
+    ).catch((err) => {
+      console.log(err);
+      return null;
+    });
 
-    const product = await Product.create({
+    if (!imagesUrl) {
+      return res.status(500).json({ message: "Error occured while adding Product" });
+    }
+
+    await Product.create({
       name,
       description,
       price: Number(price),
@@ -35,7 +40,7 @@ const createProduct = async (req, res) => {
       bestSeller: bestSeller === "true" ? true : false,
     });
 
-    res.status(201).json({ success: true, message: "Product added" });
+    res.status(201).json({ success: true, message: "Product added successfully" });
   } catch (error) {
     res.status(500).json({ message: `Internal Server error: ${error.message}` });
   }
@@ -45,14 +50,35 @@ const deleteProduct = async (req, res) => {
   try {
     const id = req.params.id;
 
-    const product = await Product.findByIdAndDelete({ _id: id });
+    const product = await Product.findById({ _id: id });
 
     if (!product) {
-      return res.status(200).json({ success: false, message: "Product not found" });
+      return res.status(400).json({ success: false, message: "Product not found" });
     }
+
+    const result = await Promise.allSettled(
+      product.image.map(async (imgUrl) => {
+        const publicId = imgUrl.split("/").at(-1).split(".").at(0);
+        console.log(imgUrl, publicId);
+        const deleted = await cloudinary.uploader.destroy(publicId, {
+          resource_type: "image",
+        });
+        console.log(deleted);
+      })
+    ).catch((err) => {
+      console.error(err);
+      return null;
+    });
+
+    if (!result) {
+      return res.status(500).json({ message: "Error occured while deleting Product" });
+    }
+
+    await Product.findByIdAndDelete({ _id: id });
 
     res.status(200).json({ success: true, product, message: "Product deleted" });
   } catch (error) {
+    console.log(error);
     res.status(500).json({ message: `Internal Server error: ${error.message}` });
   }
 };
